ViewModel - ViewModelFactory 코드 분리
